Add the base for the Navbar
In order to have a navbar shown on every page, we edit the
application.html.erb file in views/layouts. We add a partial render
action in the body and created a new folder (app/views/shared) that
contains the Navbar element. We strip it down and add basic links
between the landing and about pages.