Add our first DB model and our first user
We download bcrypt, setup our first db model to record the email and
hashed password of our users and create our first user so that the DB
is no longer empty. Through the use of the rails console we interact
with all this information and experiment. Next will be fixing some
issues our current model has, such as the fact that the email attribute
can be empty and still go through