Fix alerts not firing when form was incorrect
As it turns out, in our registrations_controller in the create action
on the else block we NEED to add "status: 400" (technically it should be
any appropiate status number in the 400 range) on our render line.

Otherwise our previous flash[:alert] line won't ever fire. Apparently it
has something to do with how the HTTP requests interact. Was a bit hard
to track down so it deserved it's own special commit