Change the main controller to User.find_by
The ".find_by" action is a better candidate for our log-in feature than
the ".find" we previously used. This is because it won't throw an app
breaking error when it can't find the ID for some reason. (a practical
example would be whenever a user is deleted from the DB but that user
still has their session open. ".find" will break in that situation.)