Create current.rb model, add signup links to _navbar
We have created a new model for a "Current" object that will store, for
every visitor, some information about their current session. So far,
it's only used to store their user and check wether they are logged in
or not. This meant a retool of our views and we took advantage of this
opportunity to retool the navbar, improve the look of the logout button
and clean up (in a very barren way, still) the look of the landing page.

As of this commit, the navbar contains links to the log in and sing up
forms, it allows us to log out when necessary and from the DB side
our app can take in new users and check users in the DB for login
purposes.